+++ demo/fastapi_app.py
@@ -17,7 +17,7 @@ language_config._attn_implementation = 'eager'
 vl_gpt = AutoModelForCausalLM.from_pretrained(model_path,
                                               language_config=language_config,
                                               trust_remote_code=True)
-vl_gpt = vl_gpt.to(torch.bfloat16).cuda()
+vl_gpt = vl_gpt.to(torch.float16).cuda()
 
 vl_chat_processor = VLChatProcessor.from_pretrained(model_path)
 tokenizer = vl_chat_processor.tokenizer
@@ -43,7 +43,7 @@ def multimodal_understanding(image_data, question, seed, top_p, temperature):
     pil_images = [Image.open(io.BytesIO(image_data))]
     prepare_inputs = vl_chat_processor(
         conversations=conversation, images=pil_images, force_batchify=True
-    ).to(cuda_device, dtype=torch.bfloat16 if torch.cuda.is_available() else torch.float16)
+    ).to(cuda_device, dtype=torch.float16 if torch.cuda.is_available() else torch.float16)
     
     inputs_embeds = vl_gpt.prepare_inputs_embeds(**prepare_inputs)
     outputs = vl_gpt.language_model.generate(
